refactor(test): move #1422's inline TelegramState tests under src/tests/
#1422 added seven tests as an inline `mod tests` in
src/channels/telegram/state.rs, against the house rule that tests live
under src/tests/. The 177 lines of test body move verbatim, checked by
diffing the block extracted from the pre-move file against the new one;
the only change is that `use super::*` becomes explicit crate paths, and
the two fold-dedupe constants the tests pin widen from private to
pub(crate) so they read the real window rather than restating it.

Count-neutral: 7,946 listed across 794 modules before and after, one
inline module out and one src/tests module in. The inline remainder
drops from 22 to 15, presets_test.rs alone, which is what the README and
TESTING.md lines now say.
