+++ src/tests/mod.rs
@@ -829,6 +829,7 @@ pub mod telegram_send_input_file_test;
 pub mod telegram_send_string_coercion_test;
 pub mod telegram_send_thread_id_override_test;
 pub mod telegram_split_message_test;
+pub mod telegram_state_test;
 pub mod telegram_status_message_test;
 pub mod telegram_stream_loop_resume_test;
 pub mod telegram_target_resolver_test;
